Fix glitches in big AnimatedSprite animations
AnimatedSprite creates an atlas texture where the frames are placed
next to each other in a row - or multiple rows, if the frames are many
and/or big enough.

When interpolation is enabled, the code must calculate not only the
coordinates for the current (A) frame, but also for the next
one. However, the code for caluclating next (B) frame was broken. It
seemed to be half way designed to operate on normalized (qreal
0.0-1.0) coordinates, while the function in reality operates on
physical, integer texture coordinates. The result was that every time
the B frame should be in a different row than the A frame, the B
coordinates came out wrong. Hence, at that point the rendering would
interpolate the A frame towards a wrong or invalid frame, causing
visual glitches.

Fix the computations to use the actual texture (sheet) size instead of
normalized coordinates, and integers instead of qreals.

Fixes: QTBUG-94765
Pick-to: 6.2 5.15
Change-Id: I5a484c196abc3189f6dce0921cad3a25ada0e584
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>