Fix Quick Palette test after qtbase change
The test tst_palette started to fail on Linux systems when updating
dependencies to qtbase
56bd1b76d2e81c58a80bf6b5d74219c6b1ab8499.

However, the test is bogus: it takes a copy of the theme's system
palette, modifies that copy (without setting it back to the theme), and
then tests that the palette of previously created Quick items is
identical to that modified copy.

There is no reason why it should be, other than because the bug fixed
in qtbase was actually breaking palette inheritance, with the resolve
mask bits set on all copies of the global palette, thus overriding the
actual style as a side effect.

There's no reason why the test would need to modify a copy of the theme
palette, or the theme palette itself, so remove those modifications.

Change-Id: Ib074882d044bdc63e28a5735343a617936bcbadf
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 05c81c5acec458b1fa294f0f9c95975fb24debaa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>