qmllint: Warn about assigning numeric types to strings
While it is generally possible to assign a numeric to a value of string
type, this is not the case for literal bindings: There, the engine's
QQmlTypeValidator notices the mismatch, and rejects the program.
Bring qmllint in line with that behavior.
Additionally, teach parseLiteralBinding to treat constant
UnaryMinusExpressions as NumberLiteral bindings.

Change-Id: Iad221e90eb1de81cabdddf5d601cc30f53ef20d3
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>