qmllint: Initial translation binding support
We will need to distinguish between script and translation bindings in
the future (to correctly assign script binding indices). Moreover, the
NOOP translation functions actually produce string literal bindings, not
translation bindings.

To unify the classification of translation bindings, the code from the
IRBuilder has been converted into a template taking a few callbacks.

This change does not add validation for translation bindings yet, but a
failing autotest for a case where it would be necesseray is already
provided.

Change-Id: Icccba98edbdcd15068188807e8622c1bd513725c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>