Deprecate QQuickFontDialog::currentFont
[ChangeLog][QtQuickDialogs] To make the font dialog's API consistent with the file and folder
dialog, we're deprecating the "currentFont" property and replacing it
with the "selectedFont" property, which previously was used as the final
selection made by the user.
By the time the user accepts the dialog, the currentFont and
selectedFont had the same value anyways. So if the application needs to
react "live" to which font the user is selecting, they can now simply
bind to the selectedFont property, and if the application only wants to
update a font when the user accepts the dialog, they can handle the
accepted signal, and read from the selectedFont property in the signal
handler.

Fixes: QTBUG-98563
Change-Id: I744bb421a466748d41e5e9c6fd6978c45601984c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>