QQuickSelectionRectangle: Don't leak the pointer handlers
Pointer Handlers reuse QObject::parent() for parentItem(). A handler
should not be instantiated in C++ with no parent, because it could leak
at shutdown if left in that state. But now we have
QQuickPointerHandler::setParentItem(), which hides the details of
adding it to ExtraData::pointerHandlers, so we no longer need to do
that explicitly here; and it doesn't seem harmful to otherwise allow
the handlers' QObject::parent to be the SelectionRectangle itself.

Change-Id: Ifcff14329d7a550c4d3749a09dc015bc112a1d28
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>