Don't intersect clipRect() with viewport if not ItemObservesViewport
We noticed after 9b62f4c27ac3fb3dc563c7f4657094c14d752bac and
9db23e0e04906cf9ea33e23fa41f34955e5e6fe0 that text in the Controls
examples was not being rendered reliably.

Before, clipRect() was always constant. After
9b62f4c27ac3fb3dc563c7f4657094c14d752bac, it started calculating the
intersection with the clipRects(s) of the viewport(s). But in these
controls, updatePaintNode() creates a QQuickDefaultClipNode instance,
using the clipRect() of either the control itself or (in the case of
TextArea) the parent Flickable, if any. These controls aren't ready yet
to deal with dynamically-changing clipRects, which demonstrates that
it's wrong for QQuickItem::clipRect() to calculate the intersection
unconditionally. An item declares its interest in that by setting the
ItemObservesViewport flag; if the flag is not set, clipRect() now
behaves as it did before. This means non-observing children of observers
also get the same old clipRect() as before; so expectedContentClipRect
in tst_QQuickItem::viewport() is rather boring now.

Change-Id: I4f1f76ccd8903c433444d2032550ee9eb92d1fd8
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>