QQmlJSScope: Avoid unnecessary lazy loading
We can amend the factory to store additional information about the type.
In particular we need its singleton status because that can be set from
the qmldir, and we need the import that was used to load it for type
resolution.

Change-Id: Ia58a393fdb72711b4b9f99a11bc594920abcad07
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit d6c7617748e8fcb472333985f75e137a0b965e04)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>