deliverSinglePointEventUntilAccepted(): skip items that aren't shown
This function is called for delivering other types of pointer events
than mouse and touch: tablet, wheel and native gesture events mainly.
It calls item->window()->mapToGlobal() so a crash is theoretically
possible in rare complex cases if pointerTargets() somehow found
items that are not in the window; so we add a null pointer check to
avoid crashing.

Fixes: QTBUG-82013
Change-Id: Id5b167a685a3ee7ae2d94338ce312930430e081b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>