Don't crash when accessing QVariant data pointer
The write method is called with a QVariant, and even though the property
might be of a QObject type, the variant might not contain a QObject. So
accessing the variant data directly via QVariant::constData and
static_cast'ing the void* to QObject* is not safe.

Add an explicit check before accessing. An alternative would be to at
least Q_ASSERT that the result of the cast and a QVariant::value call is
the same, which would then not introduce any performance penalty in
release builds. However, users use release-builds of qml and Qt tooling
to write QML, and we writing wrong QML code should not crash then
either.

To avoid converting from type ID int to QMetaType at runtime, this
cherry-pick includes a part of the refactoring done in c4addf7d0e91cd.

Include a test case that segfaults without the fix.

Fixes: QTBUG-98367
Change-Id: Ib3ae82d03c9b2df6251ee88d5bd969dd4f796a41
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit d62c216cbd09fb43302b146d5d8aca529f2b70a7)
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>