Use clipRect() again for clipping
clipRect() calculates the intersection with the viewport rect only if
ItemObservesViewport is set; but items that set that flag need to
deal with movement in the viewport by overriding transformChanged().
We modified QQuickWindowPrivate::updateDirtyNode() to construct the
QQuickDefaultClipNode using boundingRect() to avoid having it become
outdated so quickly during scrolling; but it's also unintuitive, and
contrary to what the new docs say. For each item with that flag set,
we need to mark the clip nodes dirty regardless of whether
transformChanged() does any changes to the other SG nodes, because if
clipRect() was called and then the item moves in the viewport, it will
need to be called again to redo the intersection calc.

Amends 9b62f4c27ac3fb3dc563c7f4657094c14d752bac

Change-Id: Ib9c22020540878b96369a4205b2d42bb8fd2661b
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 6a3ad603cd87071833fd45e8617a805783e76c30)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>