Make atlasing of compressed textures opt-in again
With the port to RHI of the compressed texture code for Qt 6,
automatic atlasing of compressed textures was enabled by default. It
turns out that this will many times break, depending on the specific
texture format, graphics stack and hardware driver. So we revert to
such atlasing to be opt-in.

[ChangeLog][QtQuick] Disable atlasing of compressed textures by
default. Can be enabled with QSG_ENABLE_COMPRESSED_ATLAS=1

Fixes: QTBUG-98937
Pick-to: 6.3 6.2
Change-Id: I582a4b140ccfd9f112813da6365bb7498e8ef39f
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>