TreeViewDelegate: hide indicator if it's outside the delegate
If the column is too narrow to show the indicator and the
label (in case the model depth of the row is too great), the
label will be hidden since it's elided. But the indicator will
still show, even if it's far outside the bounds of the
TreeViewDelegate.

With this patch we'll now observe the delegate's geometry, and make
sure to hide the indicator if it ends up outside the column. This is
basically a cheaper alternative than to clip the whole delegate.

Pick-to: 6.3
Change-Id: I6f8c163f6be6c44266be418c6a7ba9ea52a5faa7
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>