Warn about duplicated ids
Besides emitting a warning in case of duplicated ids, this commit also
refactors the id handling by moving it out of the generic visit
function, into a dedicated method. Moreover, internal documentation is
added to various functions in QQmlJSScopesById, and a new function to
quickly check for duplicated ids in a file is added. While those
conflicts do not necessiate an id duplication (they could be in
different components), this allows for a quick pre-check, and might be
later used to print a non-fatal warning about the potentially confusing
id reuse.

Fixes: QTBUG-99179
Change-Id: Ia99ebd17491d91d6f7747d65b3a8d15fa1f84e07
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>