TextEdit: hide cursor outside viewport if ItemObservesViewport
Given a TextEdit in a Flickable, when someone scrolls the text such
that the text cursor is completely outside the viewport, we want the
text cursor to be hidden. It is clipped if Flickable has clip: true,
which is better for "pixel-perfect" clipping, but has a performance
penalty. Anyway it still looked ridiculous in the case when clipping
is not turned on, to have a stray text cursor appearing on top of
some neighbor items, even though the text wasn't visible.

To have awareness of the viewport (such as Flickable) within which it's
shown, TextEdit has to opt-in with the ItemObservesViewport flag so that
QQuickTextEditPrivate::transformChanged() will be called; the question
is under what conditions. We already opt-in when the text is large;
otherwise we don't really want to have updatePaintNode() called so
often, unless other code opts in by setting that flag; because for short
texts, the old way is more efficient (populate all the glyphs from the
document into the scenegraph one time, and no need to update until
editing happens).

Task-number: QTBUG-83069
Pick-to: 6.3
Change-Id: Idd7cc9d2369d67b967eec50fe81964e6abe84e9c
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>