Avoid acccessing global static during destruction
We cannot access the global static through its accessor function anymore
while its dtor is running (due to a recent change in qtbase). However,
to properly keep the removal showing in the profiler, we still want
to execute removeFromCache.
As we know that we can still access the store pointer in its own
shrinkCache methods, simply pass it there to
QQuickPixmapData::removeFromCache. removeFromCache is adjusted to only
load the pointer from the global static if it is actually nullptr.

Fixes: QTBUG-99192
Pick-to: 6.3
Change-Id: Id439931874455d9bf3d48617ac108e72a22f1676
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>