Ensure that min <= max in qBound call
It is not impossible for (1.0 - visualPos) to be negative.
Which can cause the max parameter to be smaller than the min
parameter in the qBound() call.

This is a followup to c77b9bae690053b31c9db110ee5ad6d5741d4511
which partially fixes the issue, but forgot to apply the same
change to the statement before it.

I somehow managed to crash the dialogs manual test, when clicking
around randomly in the FolderDialog.
(thus reproducing it might end up being difficult)

Task-number: QTBUG-101174
Change-Id: I92b2228f1cf92d3e3d79b05d281e05c765466101
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit f2791ba6290c7d22ac9093a948e0f5bef671247b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>