CMake: Fix QNX tests to link when cross-compiling
Apparently CMake does not consider link dependencies brought via
INTERFACE_LINK_LIBRARIES as implicit dependencies to be added to
IMPORTED_LINK_DEPENDENT_LIBRARIES.

This means CMake won't add an -rpath-link flag when we expect it to,
leading to some tests failing to link when cross-compiling.

Until we figure out if our assumption is wrong or it's a bug in CMake,
fix the issue by linking explicitly to a few more dependencies
and thus force the addition of those dependencies to the
IMPORTED_LINK_DEPENDENT_LIBRARIES property of the generated
QtFooTargets.cmake files.

The issue was always there, but it got triggered recently in the CI
because 701294e3624fd76db2e71c2f620feafc59bda350 landed in qtbase
in conjunction with a submodule dependency update.

Pick-to: 6.3 dev
Task-number: QTBUG-86533
Fixes: QTBUG-99275
Change-Id: I89a87c6cca5d7ae6c04e7f672ee606444452391a
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>