CMake: Fix lookup of Python for qtdeclarative
Always use FindPython instead of find_program because the former
looks for Python in more locations (like the Windows registry).

We switch to FindPython instead of FindPythonInterp because the latter
is deprecated, and the former is available in our minimum support
CMake version (3.16).

Because find_package is called in configure.cmake that is called by a
function() call, we need to make the found interpreter a cache
variable to make it available outside of the function.

Fixes: QTBUG-99273
Change-Id: Idb6620a01b4b5764f130f9dc5780d73cd2151965
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit e0eb75b2a0578f4ef401bb6eaa8c4f320ffa2078)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>