scenegraph: Allow requesting a HDR swapchain
...with an environment variable for now, to enable development and
testing.

Setting QSG_RHI_HDR to scrgb (or extendedsrgblinear) will request an
scRGB+RGBA16F swapchain. Setting it to hdr10 requests
HDR10(PQ)+A2RGB/BGR10. Otherwise the format stays the default SDR
RGBA/BGRA8.

When something is requested and is not supported for the window, print
a helpful message in the logs with QSG_INFO=1. When supported, the logs
will contain a Creating scRGB/HDR10 swapchain message.

In multi-screen configuration it is important to remember that the
position of the window may be important, depending on the windowing
system. Just having one HDR-enabled display does not mean the
application will launch with the requested HDR format - it may only be
successful if the window is positioned on that particular screen upon
startup. When in doubt, disable all non-HDR screens and try running with
the single HDR-enabled one.

Change-Id: Icf4994d86e39d0fafb3b1321229fc7be1500b372
Reviewed-by: Andy Nichols <andy.nichols@qt.io>