scenegraph: Make the vertex/index buffer logic more robust
Be consistent in handling a failing create(). Do not try to enqueue
resource updates if we did not successfully built the buffer. Do not try
to render either.

With just a few checks we can now fully survive failing all vertex/index
buffer creation (i.e. warnings will be printed and nothing will be
rendered by Quick, but the application will continue to work). This
matches the behavior one gets when the uniform buffer creation fails,
that is already robust enough.

Pick-to: 6.3 6.2
Fixes: QTBUG-99477
Change-Id: I2ca0f3ada8a7449c5ac1346216a58b06658b5cd5
Reviewed-by: Andy Nichols <andy.nichols@qt.io>