QQuickScrollBar: ensure '0 ≤ size ≤ 1' and '0 ≤ minimumSize ≤ 1'
It doesn't make sense for these properties to be negative or larger
than 1.0.

The size property is meant to be the size of the scrollbar relative to
the scroll area size, either horizontally or vertically, and should be
in the value range 0 ≤ x ≤ 1, where 1 means that it covers the whole
scrollbar area, and 0, meaning it's invisible. This property can be set
from qml, or by QQuickFlickable with a scrollbar assigned to the
Scrollbar.vertical or Scrollbar.horizontal properties.

The minimumSize property can be set by the qml user, to force the
scrollbar to be larger than the Flickable size / contentSize ratio.
It makes sense to force the 0 ≤ x ≤ 1 constraint for this property's
value as well.

amends ef69623648f6f5ef5ab10dd9dee72f044334d53d

As an additional sanity check, I've added a qt_is_finite() check for all
qreal properties as well, just to prevent our API users from
accidentally setting a property to an invalid floating point value.

Pick-to: 6.3 6.2
Change-Id: I9f2880333483bf584986801d6b7204fea8d66e58
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>