Fix handling of QML_COMPILE_HASH
We need to check for the contents of the tag file, not its name. We need
to strip the contents before checking. We need to run git in the right
directory. We need to fail if we cannot find any QML_COMPILE_HASH.

Task-number: QTBUG-99608
Change-Id: Ic42a073b196143f8576a84e7a4531b5f2927fb68
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 0616a44e8ad0faa9c441212e281a61136c287f73)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>