QQuickTreeView: be able to expand rows that are not yet visible in the view
If the app expands a row in TreeView, that row will for a brief moment
be expanded in the proxy model, but not in the view, until the
view is polished. For that reason, when determining if a row can be
expanded or not, we should check the state of the proxy model, and
not the view.

Note: if the application needs the state of the view to reflect the
state of the proxy model immediataly after a call to expand,
forcePolish() is always possible.

Task-number: QTBUG-91374
Change-Id: I5b85e0b5c3b32d17f93409153041df3fc2437424
Reviewed-by: Andy Shaw <andy.shaw@qt.io>
(cherry picked from commit a43912865e294f2014f03eb7073ecd9ecf387623)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>