sg: Add logic in threaded loop to recognize the lack of proper vsync
Attempt to recognize that something is wrong with throttling and fall
back to system timers (i.e. do the same what we would do if there were
more than one windows, or when we knew that swap interval was set to 0
explicitly).

This aims to eliminate the well-known visual problems (i.e. too fast
advancing animations) when running with the default threaded render
loops on systems where vsync-based throttling is disfunctional. A
common example would be certain virtual machines.  Traditionally the
solution here has been to run with QSG_RENDER_LOOP=basic. This will,
in many cases, not be necessary from now on.

As with any of the existing system timer fallback logic, this does
nothing for render thread animations (the animators, such as XAnimator
and friends). Those work as before: if there is no vsync, they will
advance too fast, no solution for that (apart from not using
animators).

Another slight downside is that animations may still appear running
too fast for the first few frames because we average the results from
a number of frames before deciding something is wrong with the window.

[ChangeLog][QtQuick] The threaded render loop of Qt Quick can now
potentially recognize that the graphics stack in a particular system
is broken and does not provide vsync-based throttling. This can now
automatically trigger falling back to using regular system timers to
advance animations. (meaning one does not need to manually switch over
to the 'basic' render loop or explicitly disable vsync on the window)
Note however that this does not affect render thread animations (the
so-called animators, such as XAnimator). Those will continue to
advance based on the presentation rate no matter what.

Fixes: QTBUG-99952
Change-Id: Ied7614c9149fa3cae32d8e0aab35f614bc86a3ac
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>