SwipeView: don't cull every child item that gets added
This reverts commit a055629f43cf8589ff6d69e46b2610429aaa4167.
A SwipeView is a ListView, which means as long as the side-by-side
delegates are the right size, only one of them should occupy the area
of the SwipeView itself. The other delegates (isCurrentItem == false)
are to the sides, and might be outside the window, hidden under other
items, or clipped. So hopefully there are not users who rely on this
culling to hide them.

Fixes: QTBUG-99547
Task-number: QTBUG-51078
Task-number: QTBUG-51669
Change-Id: Ic90dc1c44b9a36dc717238b47003b4d88e91f789
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit e10de033f4c44855de7287a97e2aa651f648742e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>