Turn qmljs into app
It's tool, which means it's not built when cross-compiling. However,
it's to be desired on a target to run QML-only projects.

The earlier way to build app for a cross target was to set
QT_BUILD_TOOLS_WHEN_CROSSCOMPILING to ON.  That seems like overkill.

Task-number: QTBUG-100040
Change-Id: I64236fb00a5d69862ce0c34ef9afca092d5d0436
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 8c6b225f2aec39cfa255a183c28d8a78fa84b840)
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>