CMake: Don't cross-compile qjtest
Previously it was only excluded on ANDROID, but cross-compiled for
other platforms.

The executable should only be built when the target Qt is meant
to be used as a development SDK, to be executed on some host machine.

Because at the moment we lack a proper abstraction for that, as a work
around, we exclude building it when cross-compiling.

The executable is not called automatically by any test, it is only run
manually by a developer.

In the future, the qt_internal_add_tool call should likely be replaced
by a customized qt_internal_add_app call, installing the executable
into $prefix/libexec rather than $prefix/bin.

Amends 5a55e526382f0bb6eb647c2e6b7ce55661889990

Pick-to: 6.2 6.3
Task-number: QTBUG-92591
Task-number: QTBUG-100040
Task-number: QTBUG-100047
Change-Id: If20e73fe378acd51fe41d181078d273a7842fc81
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>