qmldom: resolve local alias properties
alias properties are resolved, until they refer to a property of an
imported type, or to a local type.
Alias of aliased properties are not officially supported, but works
in most cases. Set LocallyResolvedAlias::nAliases to non zero if such
an access happens to be able to warn about it.

Fixes: QTBUG-98276
Change-Id: I7c123d0a49d3ae283fc4c06e449f67b1d89e72c9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>